Fix tests and other fixes
