+++ sunpy/net/dataretriever/client.py
@@ -7,8 +7,7 @@ import datetime
 from sunpy.time import TimeRange
 from sunpy.util import print_table
 
-from sunpy.net.download import Downloader
-from sunpy.net.vso.vso import Results
+from sunpy.net.download import Downloader, Results
 from sunpy.net.vso.attrs import Time
 
 __all__ = ['QueryResponse', 'GenericClient']
