+++ sunpy/net/dataretriever/tests/test_goes_ud.py
@@ -2,8 +2,8 @@ import pytest
 
 from sunpy.time.timerange import TimeRange
 from sunpy.net.vso.attrs import Time,Instrument
-from sunpy.net.unifieddownloader.client import QueryResponse
-import sunpy.net.unifieddownloader.sources.goes as goes
+from sunpy.net.dataretriever.client import QueryResponse
+import sunpy.net.dataretriever.sources.goes as goes
 
 LCClient = goes.GOESClient()
 
@@ -45,7 +45,7 @@ def test_query():
     assert len(qr1) == 1
     assert qr1.time_range()[0] == '2012/08/09'
     assert qr1.time_range()[1] == '2012/08/10'
-    
+
 
 @pytest.mark.online
 @pytest.mark.parametrize("time,instrument",
